+++ sdk/search/search/src/index.ts
@@ -55,7 +55,6 @@ export {
   SimpleDataType,
   ComplexDataType
 } from "./serviceModels";
-export { SearchApiKeyCredential } from "./searchApiKeyCredential";
 export { default as GeographyPoint } from "./geographyPoint";
 export { odata } from "./odata";
 export { IndexDocumentsBatch } from "./indexDocumentsBatch";
@@ -144,3 +143,4 @@ export {
   TokenInfo,
   AccessCondition
 } from "./generated/service/models";
+export { AzureKeyCredential } from "@azure/core-auth";