+++ sdk/search/search/CHANGELOG.md
@@ -6,6 +6,7 @@
 - [Breaking] `indexDocuments` on `SearchIndexClient` now takes an `IndexDocumentsBatch` object instead of a raw action array. This new type helps compose an array of actions to be performed on the index.
 - [Breaking] In `SearchIndexClient`, removed options `mergeIfExists` and `uploadIfNotExists` on `uploadDocuments` and `mergeDocuments` and replaced them with new helper `mergeOrUploadDocuments`.
 - The type `IndexAction` was renamed to `IndexDocumentsAction`.
+- [Breaking] Removed `SearchApiKeyCredential` and replaced with `AzureKeyCredential`.
 
 ## 11.0.0-preview.1 (2020-03-09)